Hi, I need help please.
The test is not running successfully but I cannot understand why there is all these issues.  I must admit I copied most of this from the homework slides as I tend to do better when seeing the result and emulating.  But I dont know why its not correctly running or why all the errors.  Even in places where it states there is a syntax error and I dont see one.  Or, the other errors.  Could you help explain please?